requirements.txt
# Core dependencies
supabase==1.0.4
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
